feat(condo): DOMA-9722 updated mailing agreement
